Initial Commit
Functional exp searcher, but EXTREMELY slow (on the order of a minute on debug builds)
Unclear how much tuning is needed for the pattern, would be great if we could query a full page for zero/non-zero properties quickly, or find a closer reference page
Also need to add tests, make the scanning truly async, and add the actual logic for doing things based off of actions.
Should split into files to do work elsewhere and have the main loop delegate accordingly.